[MIG] currency_rate_date_check: Migration to 10.0
