--- currency_rate_date_check/i18n/ro.po
@@ -1,57 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# * currency_rate_date_check
-# 
-# Translators:
-msgid ""
-msgstr ""
-"Project-Id-Version: account-financial-tools (8.0)\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2016-12-23 23:21+0000\n"
-"PO-Revision-Date: 2015-06-03 15:55+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: Romanian (http://www.transifex.com/oca/OCA-account-financial-tools-8-0/language/ro/)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Language: ro\n"
-"Plural-Forms: nplurals=3; plural=(n==1?0:(((n%100>19)||((n%100==0)&&(n!=0)))?2:1));\n"
-
-#. module: currency_rate_date_check
-#: model:ir.model,name:currency_rate_date_check.model_res_company
-msgid "Companies"
-msgstr "Companii"
-
-#. module: currency_rate_date_check
-#: model:ir.model,name:currency_rate_date_check.model_res_currency
-msgid "Currency"
-msgstr ""
-
-#. module: currency_rate_date_check
-#: field:res.company,currency_rate_max_delta:0
-msgid "Max Time Delta in Days for Currency Rates"
-msgstr ""
-
-#. module: currency_rate_date_check
-#: sql_constraint:res.company:0
-msgid ""
-"The value of the field 'Max Time Delta in Days for Currency Rates' must be "
-"positive or 0."
-msgstr ""
-
-#. module: currency_rate_date_check
-#: help:res.company,currency_rate_max_delta:0
-msgid ""
-"This is the maximum interval in days between the date associated with the "
-"amount to convert and the date of the nearest currency rate available in "
-"Odoo."
-msgstr ""
-
-#. module: currency_rate_date_check
-#: code:addons/currency_rate_date_check/currency_rate_date_check.py:83
-#, python-format
-msgid ""
-"You are requesting a rate conversion on %s for currency %s but the nearest "
-"rate before that date is dated %s and the maximum currency rate time delta "
-"for your company is %s days"
-msgstr ""
