+++ currency_rate_date_check/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- encoding: utf-8 -*-
+
+from . import company
+from . import currency_rate_date_check
