+++ currency_rate_date_check/views/company_view.xml
